Export Notes + Setup Database for Quiz + Bugfixing Diary notes
Now you can export all your diary's notes and your weekly notes.
Did some initial setup for when the quiz will be ready.
Fixed a bug in that caused the note's description with line brakes in it to bug the front-end.

Pulling this change will reset your database, be careful